+++ bold-pagos-en-linea/include/BoldMenuAdmin.php
@@ -50,17 +50,17 @@ class BoldMenuAdmin
     private function addAssetsAdmin(): void
     {
         wp_register_script( 'woocommerce_bold_admin_panel_js', plugins_url( '/../assets/js/admin-panel.js', __FILE__ ),
-            array('bold-assets-js', 'jquery'), '3.3.3', true );
+            array('bold-assets-js', 'jquery'), '3.4.0', true );
         wp_enqueue_script( 'woocommerce_bold_admin_panel_js' );
-        wp_enqueue_style( 'woocommerce_bold_admin_panel_css', plugin_dir_url( __FILE__ ) . '../assets/css/bold_admin_panel_style.css', false, '3.3.3', 'all' );
+        wp_enqueue_style( 'woocommerce_bold_admin_panel_css', plugin_dir_url( __FILE__ ) . '../assets/css/bold_admin_panel_style.css', false, '3.4.0', 'all' );
 
-        wp_register_script( 'woocommerce_bold_icons-dark-ui', BoldConstants::URL_CHECKOUT.'/library/ui-kit.js?layout=vertical&type=slider&target=bold-config-dark-icons', null, '3.3.3', true );
+        wp_register_script( 'woocommerce_bold_icons-dark-ui', BoldConstants::URL_CHECKOUT.'/library/ui-kit.js?layout=vertical&type=slider&target=bold-config-dark-icons', null, '3.4.0', true );
         wp_enqueue_script( 'woocommerce_bold_icons-dark-ui' );
-        wp_register_script( 'woocommerce_bold_icons-light-ui', BoldConstants::URL_CHECKOUT.'/library/ui-kit.js?layout=vertical&type=slider&theme=dark&target=bold-config-light-icons', null, '3.3.3', true );
+        wp_register_script( 'woocommerce_bold_icons-light-ui', BoldConstants::URL_CHECKOUT.'/library/ui-kit.js?layout=vertical&type=slider&theme=dark&target=bold-config-light-icons', null, '3.4.0', true );
         wp_enqueue_script( 'woocommerce_bold_icons-light-ui' );
 
         wp_enqueue_media();
-        wp_enqueue_script('woocommerce_bold_media_uploader', plugins_url( '/../assets/js/bold-media-uploader.js', __FILE__ ), ['jquery', 'wp-i18n'], '3.3.3', true);
+        wp_enqueue_script('woocommerce_bold_media_uploader', plugins_url( '/../assets/js/bold-media-uploader.js', __FILE__ ), ['jquery', 'wp-i18n'], '3.4.0', true);
         wp_localize_script('woocommerce_bold_media_uploader', 'BoldPlugin', ['pluginUrl' => plugin_dir_url(__DIR__)]);
     }
 
@@ -95,7 +95,7 @@ class BoldMenuAdmin
         }
 
         if (isset($_SERVER["REQUEST_METHOD"]) && sanitize_text_field(wp_unslash($_SERVER["REQUEST_METHOD"])) === "POST") {
-            if ( empty( $_POST ) || ! check_admin_referer( 'bold_update_settings', 'bold_settings_form_nonce' ) ) {
+            if ( empty( $_POST ) || empty( $_POST['bold_settings_form_nonce'] ) || ! wp_verify_nonce( sanitize_text_field( wp_unslash( $_POST['bold_settings_form_nonce'] ) ), 'bold_update_settings' ) ) {
                 $message_error = esc_html__('No se pudo guardar la información, no paso la verificación de seguridad. Intente de nuevo.', 'bold-pagos-en-linea');
                 $error_notice = new BoldNotice('warning', $message_error);
                 $error_notice_script = "<script>document.addEventListener('DOMContentLoaded', function() {if(window.Notiflix) window.Notiflix.Notify.warning('".$message_error."',{zindex: 99999});});</script>";